rename all calls to `database_migrator` to `data_migrator`
