+++ cea/scripts.yml
@@ -276,13 +276,13 @@ Optimization:
       - [PV_totals]
 
 Utilities:
-  - name: database-migrator
-    label: Database migrator
+  - name: data-migrator
+    label: Data migrator
     description: "Migrate older scenario inputs to the latest version.
                   Please consider running this for scenario inputs created before v3.22.
                   NOTE: This cannot be undone - save a copy of the scenario first."
     interfaces: [cli, dashboard]
-    module: cea.datamanagement.database_migrator
+    module: cea.datamanagement.data_migrator
     parameters: ['general:scenario']
     input-files:
       - [get_database_construction_standards]
